style & fix: format codes & underline occurrence other than frame
The occurrences were framed, but it sometimes cost a long time which is
perceptible by the users. What's worse, it sometimes needs many undos,
which is described in Megaxela/QCodeEditor#20.

After this change, the occurrences are underlined other than framed,
and the words are find by whole-word rule. This is fast and doesn't
need any undos.

It may be possible to use frames which is more beatiful without being
slow or modifying the text, but I'm just lazy to implement it.