fix & refactor: Better auto indent
1. Replace the confusing code with a much more concise one.

2. Works with mixed spaces and tabs.

For example, suppose the tab length is four, `  \t  \t ` is replaced by
either `             ` or `\t\t\t` in the old version, but neither of
them is correct. The new version will keep `  \t  \t ` without any
modifcation, and the tab replace function only works for insert tabs
manually or the extra indent for '{'.