fix: Fix wrong toggle comments, indent, etc. when text wrap is on
This is reported at cpeditor/cpeditor#213.

Qt uses "block" for a real line, and "line" for a line in the GUI.
They are the same when text wrap is off, but different when text
warp is on.

To fix this, "Up" should be replaced by "PreviousBlock", "Down"
should be replaced by "NextBlock", "StartOfLine" should be replaced
by  "StartOfBlock", "EndOfLine" should be replaced by "EndOfBlock".