Fix(clearSquiggle failing on some cases)
In some cases the clear function failed because the stored start and
stop position of underline changes while not being updated in editor.
Now we select all document and apply default textFormat to clear the
underlines. It is much faster and better. squiggle Function has been
improved too, it now handles all (Valid) ranges that LSP will send to
highlight.